Refactor of all components
