+++ frontend/components/Team/team.module.css
@@ -10,11 +10,11 @@
   color: #7b7c7b;
   margin: auto;
   text-align: center;
-  padding: 50px;
+  padding: 30px 50px 50px 50px;
 }
 
 .main_container {
-  background: url("/images/home_hero/bckgrnd.png") no-repeat center;
+  background: url("/images/ourteam/bckgrnd.png") no-repeat center;
   width: 100vw;
 }
