+++ frontend/pages/_app.jsx
@@ -1,6 +1,6 @@
 import "../styles/globals.css";
 import "../styles/swiper.min.css";
-import '../styles/navigation.min.css'
+import "../styles/navigation.min.css";
 import { Provider } from "react-redux";
 import { useStore } from "../store";
