+++ frontend/pages/index.jsx
@@ -1,10 +1,12 @@
+import { categoriesApi } from "../store/categories/action";
+import { useSelector } from "react-redux";
 import { useState } from "react";
+import { withRedux } from "../hof/withRedux";
 import NavBar from "../components/NavBar";
 import SubHeader from "../components/SubHeader";
 import Popular from "../components/Popular";
-import { withRedux } from "../hof/withRedux";
-import { categoriesApi } from "../store/categories/action";
-import { useSelector } from "react-redux";
+import Team from "../components/Team";
+import Contact from "../components/Contact";
 
 export default function Main() {
   const { data: categories } = useSelector(categoriesApi.endpoints.getAllCategories.select());
@@ -16,6 +18,8 @@ export default function Main() {
       <NavBar isOpened={isOpened} setIsOpened={setIsOpened} />
       <SubHeader isOpened={isOpened} />
       <Popular />
+      <Team />
+      <Contact />
     </>
   );
 }
