+++ frontend/public/navBar.module.css
@@ -1,0 +1,146 @@
+.top_nav_bar_block,
+.humburger_menu_btn {
+  display: none;
+}
+
+.wrapper_main_head {
+  margin: auto;
+  background-color: #000b2d;
+  height: 100px;
+  display: flex;
+  align-items: center;
+  font-family: "Montserrat-Light", sans-serif;
+  font-size: 20px;
+  color: #eaeaea;
+}
+
+.main_block_container {
+  width: 1920px;
+  margin: 0 auto;
+  padding: 0 180px 0 160px;
+  justify-content: space-between;
+  display: flex;
+  align-items: center;
+}
+
+.humburger_menu_btn {
+  background-image: url(../../public/images/menu.svg);
+  width: 40px;
+  height: 40px;
+  background-size: contain;
+  border: 0;
+  background-color: #000b2d;
+  cursor: pointer;
+  z-index: 1;
+}
+
+.top_search_block {
+  font-size: 15px;
+}
+
+.top_search_block > a {
+  margin-right: 1rem;
+}
+
+.search_input_container,
+.search_input_container > input {
+  cursor: pointer;
+}
+
+.search_input_container > input {
+  height: 38px;
+  width: 240px;
+}
+
+.search_input_container::after {
+  content: " ";
+  right: 83px;
+  position: relative;
+  border-left: 2px solid;
+  color: #193f65;
+}
+
+.search_input_container > svg {
+  font-size: 15px;
+  position: relative;
+  right: 55px;
+}
+
+.top_bar_search_input {
+  background-color: #000b2d;
+  border: solid 2px #193f65;
+  margin-right: 30px;
+  outline: none;
+  color: #eaeaea;
+  font-size: 16px;
+  letter-spacing: 0.5px;
+}
+
+.color_search {
+  color: #193f65;
+}
+
+.top_bar_user_icon {
+  font-size: 23px;
+  padding-left: 70px;
+  background-image: url(../../public/images/header_nav/icon_user.png);
+  background-repeat: no-repeat;
+  background-position: center;
+}
+
+@media screen and (max-width: 1660px) {
+  .top_nav_bar_block {
+    font-size: 17px;
+  }
+
+  .main_block_container {
+    padding: 0 100px 0 80px;
+  }
+}
+
+@media screen and (max-width: 1494px) {
+  .main_block_container {
+    padding: 0 65px;
+  }
+}
+
+@media screen and (max-width: 1324px) {
+  .humburger_menu_btn {
+    display: block;
+  }
+
+  .search_input_container::after {
+    right: 40px;
+  }
+
+  .top_bar_user_icon,
+  .wrapper_links {
+    display: none;
+  }
+}
+
+@media screen and (max-width: 576px) {
+  .main_block_container {
+    padding: 0;
+    padding-left: 25px;
+  }
+
+  .humburger_menu_btn {
+    max-width: 30px;
+    max-height: 30px;
+  }
+
+  .search_input_container > svg {
+    right: 40px;
+  }
+
+  .search_input_container::after {
+    right: 33px;
+  }
+
+  .top_bar_search_input {
+    max-width: 200px;
+    max-height: 35px;
+    margin-right: 15px;
+  }
+}
